requirements.txt
aiofiles==23.2.1
altair==5.3.0
annotated-types==0.7.0
anyio==4.4.0
argcomplete==1.10.3
attrs==23.2.0
beautifulsoup4==4.8.2
bs4==0.0.2
cassandra-driver==3.29.1
certifi==2024.7.4
chardet==3.0.4
charset-normalizer==3.3.2
clang==17.0.6
click==8.1.7
compressed-rtf==1.0.6
contourpy==1.2.1
cycler==0.12.1
diskcache==5.6.3
dnspython==2.6.1
docx2txt==0.8
ebcdic==1.1.1
EbookLib==0.18
email_validator==2.2.0
exceptiongroup==1.2.2
extract-msg==0.28.7
fastapi==0.111.1
fastapi-cli==0.0.4
ffmpy==0.3.2
filelock==3.13.1
fonttools==4.53.1
fsspec==2024.2.0
geomet==0.2.1.post1
gradio==4.38.1
gradio_client==1.1.0
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.24.0
idna==3.7
IMAPClient==2.1.0
importlib_resources==6.4.0
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
llama_cpp_python==0.2.75
lxml==5.2.2
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.1
mdurl==0.1.2
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.4
olefile==0.47
orjson==3.10.6
packaging==24.1
pandas==2.2.2
pdfminer.six==20191110
pillow==10.4.0
pycryptodome==3.20.0
pydantic==2.8.2
pydantic_core==2.20.1
pydub==0.25.1
Pygments==2.18.0
pyparsing==3.1.2
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
python-pptx==0.6.23
pytz==2024.1
PyYAML==6.0.1
referencing==0.35.1
regex==2024.5.15
requests==2.32.3
rich==13.7.1
rpds-py==0.19.0
ruff==0.5.4
safetensors==0.4.3
scikit-learn==1.5.1
scipy==1.14.0
semantic-version==2.10.0
sentence-transformers==3.0.1
shellingham==1.5.4
six==1.12.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
SpeechRecognition==3.8.1
starlette==0.37.2
sympy==1.12
textract==1.6.5
threadpoolctl==3.5.0
tokenizers==0.19.1
tomlkit==0.12.0
toolz==0.12.1
tqdm==4.66.4
transformers==4.42.4
typer==0.12.3
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
urllib3==2.2.2
uvicorn==0.30.3
uvloop==0.19.0
watchfiles==0.22.0
websockets==11.0.3
xlrd==1.2.0
XlsxWriter==3.2.0
